refactor: se movio el componente PasswordConfirmationControl
