+++ frontend/src/components/custom-controls/PasswordConfirmControl.tsx
